+++ zcash_keys/CHANGELOG.md
@@ -65,6 +65,7 @@ The entries below are relative to the `zcash_client_backend` crate as of
   - `UnifiedAddressRequest`
 - A new `orchard` feature flag has been added to make it possible to
   build client code without `orchard` dependendencies.
+- `zcash_keys::address::Address::to_zcash_address`
 
 ### Changed
 - The following methods and enum variants have been placed behind an `orchard`